Change hibernate property to utc
