+++ src/main/java/fi/hsl/transitlog/hfp/configuration/DatabaseConfiguration.java
@@ -57,7 +57,7 @@ public class DatabaseConfiguration {
         properties.put("hibernate.order_inserts", true);
         properties.put("hibernate.jdbc.batch_size", 5000);
         properties.put("hibernate.order_updates", true);
-        properties.put("hibernate.jdbc.time_zone", true);
+        properties.put("hibernate.jdbc.time_zone", "UTC");
         properties.put("hibernate.jdbc.batch_versioned_data", true);
         properties.put("hibernate.show_sql", env.getProperty("hibernate.show_sql"));
         properties.put("hibernate.format_sql", env.getProperty("hibernate.format_sql"));
